init:
This script uses the dicom_parser crate to parse DICOM files and checks each file in the directory against the DICOM file format. If a file is found to be a valid DICOM file, the script prints a message to the console. If an error occurs during parsing or reading of a file, an error message is printed to the console.